==> code/preprocessing/01_ProcessCaseData.R <==
library(tidyverse)
library(sf)
library(lubridate)
library(scales)
library(patchwork)

#-------------------- Process Hantavirus Data -------------------- 

#----- AMERICAS -----
### USA ###
# data source: https://www.cdc.gov/hantavirus/surveillance/index.html
usa = read.csv('data/human/raw/hanta/USA_hantavirus_monthly.csv', skip=1)
usa = usa[1:nrow(usa)-1,] # remove last row, contains citation info
usa$country = 'USA'

# State recorded as FIPS - need to convert to state name
# FIPS codes downloaded from: 
fips = read.csv('data/location_codes/usa_state_fips.csv')
fips = fips%>%
  rename(StateFIPS = FIPS)
usa = left_join(usa, fips)

usa = usa%>%
  # remove undefined cases
  subset(IllnessOnsetDate != 'Before 1993-undefined')%>%
  rename(adm1 = state, 
         date = IllnessOnsetDate)%>%
  mutate(year = year(my(date)),
         month = month(my(date)),
         num_cases = 1)

# remove unknown date
usa = usa%>%
  subset(!is.na(year))

# now aggregate
usa = usa%>%
  group_by(country, adm1, year, month)%>%
  summarize(num_cases = sum(num_cases))%>%
  ungroup()

usa$disease = 'HPS'
usa$continent = 'America'

write.csv(usa, 'data/human/processed/hanta/US_hanta.csv', row.names = F)

### Brazil ### 
br = read.csv('data/human/raw/hanta/Brazil_hantavirus_monthly.csv')
br$country = 'Brazil'
br$disease = 'HPS'

# add state names
sf = st_read('data/location_codes/bra_shp/Brazil_shp_harm_2022.shp')
sf = sf%>%
  rename(adm1 = nam_stt, adm2 = name_mn)%>%
  select(c(adm1, IBGE6))%>%
  st_drop_geometry()

br$IBGE6 = as.character(br$IBGE6)
br = left_join(br, sf, by='IBGE6')
br = br%>%
  select(c(country, adm1, adm2, year, month, num_cases))

br$continent = 'America'

# remove not reported or outside brazil
br = br%>%
  subset(!is.na(adm1))

# convert adm1 to english
names = read.csv('data/location_codes/conversions/brazil_adm1.csv')

br = br%>%
  rename(adm1_case = adm1)

br = left_join(br, names)

# check this is 0
nas = br%>%
  subset(is.na(adm1))%>%
  select(c(adm1_case, adm1))%>%
  distinct()

br = br%>%
  select(-c(adm1_case))
  
write.csv(br, 'data/human/processed/hanta/BR_hanta.csv', row.names = F)

### CHILE ###
ch = read.csv('data/human/raw/hanta/Chile_hantavirus.csv')
ch$disease = 'HPS'
ch$country = 'Chile'
ch = ch%>%
  rename(adm1_notification = adm1,
         adm1 = region_probable_contagion)%>%
  mutate(date_first_symptoms = as.Date(date_first_symptoms, format="%d/%m/%Y"),
         month=month(date_first_symptoms),
         year = year(date_first_symptoms),
         isoweek = isoweek(date_first_symptoms))%>%
  rename(date = date_first_symptoms)

ch = ch%>%
  group_by(year, month, isoweek, date, country, adm1, adm3, disease)%>%
  summarize(num_cases = n())

# add in adm2
sf = st_read('data/location_codes/chl_shp/chl_admbndp_admALL_bcn_itos_20211008.shp')
sf = sf%>%
  rename(adm3 = ADM3_ES,
         adm2 = ADM2_ES,
         adm1 = ADM1_ES)%>%
  select(c(adm2, adm3))%>%
  st_drop_geometry()

# one has english spelling, other spanish
ch$adm3_es = gtranslate::translate(ch$adm3, from="en", to="es")
ch = ch%>%
  rename(adm3_eng = adm3,
         adm3 = adm3_es)%>%
  mutate(adm3 = str_to_title(adm3))

ch = left_join(ch, sf)

# some missing but will sort later
nas = ch%>%
  subset(is.na(adm2))%>%
  distinct(adm3)

ch$continent = 'America'

# fix some naming issues
ch = ch%>%
  subset(adm1 != 'No se pudo determinar ')

ch$adm1 = ifelse(ch$adm1 == 'SEREMI Del Maule', 'Región del Maule', ch$adm1)
ch$adm1 = ifelse(ch$adm1 == 'SEREMI De Los Rios', 'Región de Los Ríos', ch$adm1)
ch$adm1 = ifelse(ch$adm1 == 'SEREMI Del Nuble', 'Región del Ñuble', ch$adm1)
ch$adm1 = ifelse(ch$adm1 == 'Región de la Araucanía', 'Región De La Araucania', ch$adm1)

ch = read.csv('data/human/processed/hanta/CH_hanta.csv')

ch=ch%>%select(-c(adm1))

names = read.csv('data/location_codes/conversions/chile_adm1.csv')

ch = ch%>%
  rename(adm1_case=adm1)

ch = left_join(ch, names)

# check this is 0
nas = ch%>%
  subset(is.na(adm1))%>%
  select(c(adm1_case, adm1))%>%
  distinct()

write.csv(ch, 'data/human/processed/hanta/CH_hanta.csv', row.names = F)

#----- EUROPE ----- 
### Sweden ### 
sw = read.csv('data/human/raw/hanta/Sweden_Hantavirus.csv')
sw = sw%>%
  mutate(across(January:December, ~stringr::str_split_i(., '/', 1)))%>%
  rename(adm1 = County.Regions, year = Year)
sw$country = 'Sweden'
sw = sw%>%
  tidyr::pivot_longer(January:December, names_to = "month", values_to = "num_cases")
sw  = sw%>%
  mutate(month = ifelse(month == 'Sept', 'September', month),
         num_cases = stringr::str_trim(num_cases), # remove white space
         date = as.Date(paste0(year, '-', month, '-01'), format = '%Y-%B-%d'),
         month = month(date)) 

sw$num_cases = as.integer(sw$num_cases)
sw = sw%>%
  subset(num_cases > -1) # remove cases not measured yet
sw$disease = 'HFRS'
sw$continent = 'Europe'

sw = sw%>%
  rename(adm1_case = adm1)

sw = sw%>%
  select(-c(adm1_sw, adm1_sw2, adm1))

names = read.csv('data/location_codes/conversions/sweden_adm1.csv')
sw = left_join(sw, names, by=c('adm1_case', 'country'))

# check this is 0
nas = sw%>%
  subset(is.na(adm1))%>%
  select(c(adm1_case, adm1))%>%
  distinct()

write.csv(sw, 'data/human/processed/hanta/SE_hanta.csv', row.names = F)

### Germany ### 
de_list = lapply(list.files('data/human/raw/hanta/germany', full.names = T), function(x){
    y = read.csv(x)
    y = y%>%
      tidyr::pivot_longer(-Week.of.notification, names_to = 'region', values_to = 'num_cases')
    y$year = str_split_i(str_split_i(x, '/', -1), '\\.', 1)
    return(y)
  }
)

de = do.call(rbind, de_list)
de = de%>%
  mutate(num_cases = ifelse(is.na(num_cases), 0, num_cases),
         date = ISOweek::ISOweek2date(paste0(year, '-W', sprintf("%02d", Week.of.notification), '-', "1")),
         month = month(date)
  )%>%
  rename(isoweek = Week.of.notification)

# change region names
de = de%>%
  mutate(adm2 = ifelse(grepl('County', region), str_split_i(region, '\\.County', 1),
                         ifelse(grepl('City', region), str_split_i(region, 'City.of.', 2),
                                region)),
         adm2 = str_replace_all(adm2, '\\.', ' ')
  )


# need to get adm2 and adm1 names
# load shapefile
sf = st_read('~/Projects/RBDML/data/location_codes/ger_shp/DEU_adm3.shp')
sf = sf%>%
  rename(adm3 = NAME_3,
         adm2 = NAME_2,
         adm1 = NAME_1)%>%
  select(c(adm1, adm2))%>%
  mutate(adm2 = str_replace_all(adm2, '-', ' '))%>%
  st_drop_geometry()%>%
  distinct()

# fixes
de$adm2 = ifelse(de$adm2 == 'Sankt Wendel', 'St. Wendel', de$adm2)
de$adm2 = str_replace_all(de$adm2, " i d ", " in der ")
de$adm2 = str_replace_all(de$adm2, " i ", " im ")
de$adm2 = str_replace_all(de$adm2, " a d ", " an der ")
de$adm2 = str_replace_all(de$adm2, " a ", " am ")
de$adm2 = str_replace_all(de$adm2, "OPf ", "Oberpfalz")
de$adm2 = str_replace_all(de$adm2, "Greater Aachen", "Städteregion Aachen")
de$adm2 = str_replace_all(de$adm2, "Haale", "Haale Saale")
de$adm2 = str_replace_all(de$adm2, "Kempten", "Kempten  Allgäu")
de$adm2 = str_replace_all(de$adm2, "Lindau", "Lindau  Bodensee")
de$adm2 = str_replace_all(de$adm2, "Greater Hannover", "Region Hannover")
de$adm2 = str_replace_all(de$adm2, "Altenkirchen", "Altenkirchen  Westerwald ")
de$adm2 = str_replace_all(de$adm2, "Altenkirchen", "Frankenthal  Pfalz ")
de$adm2 = str_replace_all(de$adm2, "Ludwigshafen", "Ludwigshafen am Rhein")
de$adm2 = str_replace_all(de$adm2, "Stadtverband Saarbrücken", "Regionalverband Saarbrücken")
de$adm2 = str_replace_all(de$adm2, "Bitburg Prüm", "Eifelkreis Bitburg Prüm")
de$adm2 = str_replace_all(de$adm2, "Frankenthal  Pfalz   Westerwald ", "Frankenthal")
de$adm2 = str_replace_all(de$adm2, "Kempten  Allgäu", "Kempten")
de$adm2 = str_replace_all(de$adm2, "Lindau  Bodensee", "Lindau")

de= de%>%
  mutate(adm2 = ifelse(str_split_i(de$adm2, ' ', 1) == 'Berlin', 'Berlin', adm2))%>%
  subset(adm2 != 'Unknown')

# remove paranthesis
sf$adm2 = str_replace_all(sf$adm2, '\\(', ' ')
sf$adm2 = str_replace_all(sf$adm2, '\\)', ' ')
sf$adm2 = str_replace_all(sf$adm2, "Halle  Saale ", "Halle")
sf$adm2 = str_replace_all(sf$adm2, "Kempten  Allgäu ", "Kempten")
sf$adm2 = str_replace_all(sf$adm2, "Lindau  Bodensee ", "Lindau")
sf$adm2 = str_replace_all(sf$adm2, "Frankenthal  Pfalz ", "Frankenthal")

de = left_join(de, sf)

# should be 0
nas = de%>%
  select(c(adm1, adm2))%>%
  distinct()%>%
  subset(is.na(adm1))

de$country = 'Germany'

write.csv(de, 'data/human/processed/hanta/DE_hanta.csv', row.names = F)

### FINLAND ###
fi = read.csv('data/human/raw/hanta/Finland_hantavirus.csv')
fi = fi%>%
  mutate(num_cases = ifelse(is.na(num_cases), 0, num_cases))

# region names are in Swedish fyi
fi = fi%>%
  subset(Region != "Hela landet")%>% # means throughout the country
  rename(adm1 = Region,
         date = Date)%>%
  mutate(date = as.Date(paste0(date, "-01"), format="%Y-%m-%d"))

# get year and month
fi = fi%>%
  mutate(month = month(date),
         year = year(date),
         country = 'Finland')

fi$continent = 'Europe'  
fi$disease = 'HFRS'

fi = fi%>%
  rename(adm1_case = adm1)

# add names
names = read.csv('data/location_codes/conversions/finland_adm1.csv')

fi = left_join(fi, names)

write.csv(fi, 'data/human/processed/hanta/FI_hanta.csv', row.names = F)

# checked that they each match EU numbers before this
# maybe input test though
### EUROPE DATA ###
eu = read.csv('data/human/raw/hanta/Europe_hantavirus.csv')
eu = eu%>%
  rename(country = RegionName,
         num_cases = NumValue)%>%
  mutate(year = as.numeric(str_split_i(Time, '-', 1)),
         month = as.numeric(str_split_i(Time, '-', 2)),
         num_cases = as.numeric(num_cases),
         disease = 'HFRS',
         continent = 'Europe')%>%
  select(c(year, month, continent, country, num_cases, disease))

# remove NAs (no data reported)
eu = eu%>%
  subset(!is.na(num_cases))

# remove sweden
eu = eu%>%
  subset(country != 'Sweden')%>%
  subset(country != 'Germany')%>%
  subset(country != 'Finland')

#eu = do.call(rbind, c(eu, de, sw, fi))

eu$adm1 = 'all'

# remove repeated dates
#eu = eu%>%distinct()

write.csv(eu, 'data/human/processed/hanta/EU_hanta.csv', row.names = F)

#----- ASIA ----- 

### CHINA ###
dir = 'data/human/raw/hanta/china/'
cn = lapply(list.files(dir), function(x){
   y = readxl::read_excel(paste0(dir, x))
   y$year = str_split_i(str_split_i(x, '-', -1), '\\.', 1)
   y$month = str_split_i(str_split_i(x, '-', 1), '\\.', 1)
   return(y)
})

cn = do.call(rbind, cn)
cn$month = as.numeric(cn$month)
cn$year = as.numeric(cn$year)
cn$country = 'China'

t = cn%>%
  subset(year == 2023)
cn = cn%>%
  subset(Diseases == 'HFRS' | Diseases == 'Epidemic hemorrhagic fever')


cn$disease = 'HFRS'
cn$num_cases = as.numeric(gsub(",","",cn$Cases))

cn = cn%>%
  select(c(year, month, country, num_cases, disease))

cn2 = readxl::read_excel('data/human/raw/hanta/China_Hantavirus_2019.xlsx', skip=1)
cn2$disease = 'Hantavirus'
cn2$country = 'China'
cn2 = cn2%>%rename(num_cases = cases)

cn = rbind(cn, cn2)

# add NAS for break line
na = data.frame(year=2019,
                month =c(1:10),
                num_cases=NA,
                country='China',
                disease ='HFRS')

cn = rbind(cn, na)

cn = read.csv('data/human/processed/hanta/CN_hanta.csv')
cn$adm1 = 'all'
cn$continent = 'Asia'
write.csv(cn, 'data/human/processed/hanta/CN_hanta.csv', row.names = F)

#-------------------- Combine Hantavirus Data -------------------- 
df = do.call(bind_rows, 
             lapply(list.files('data/human/processed/hanta', full.names = T), 
                    read.csv))
# fix germany
df = df%>%
  mutate(continent = ifelse(country == 'Germany', 'Europe', continent))
df$adm1 = ifelse(df$adm1 == 'Baden-Württemberg', 'Baden-Wuerttemberg', df$adm1)
df$adm1 = ifelse(df$adm1 == 'Thüringen', 'Thueringen', df$adm1)
df$adm1 = ifelse(df$adm1 == 'Nuble', 'Biobio', df$adm1)
df$continent = ifelse(df$country == 'Germany', 'Europe', df$continent)

eu_removals = df%>%
  subset(continent == 'Europe')%>%
  group_by(country)%>%
  summarize(cases = sum(num_cases))%>%
  subset(cases < 300)

eu_removals = unique(eu_removals$country)
eu_removals = c(eu_removals, 'EU', 'EU/EEA')

df = df%>%
  subset(!country %in% eu_removals)

# # create country factor ordering
# order = df%>%
#   select(c(continent, country))%>%
#   distinct()%>%
#   arrange(continent)
# 
# df$country = factor(df$country, order$country)

df = df%>%
  select(c(country, adm1, year, month, num_cases, continent))

df$disease = 'Hantavirus'
#-------------------- Combine lassa and hanta  -------------------- 
las = read.csv('data/human/raw/lassa/lassafever_confirmed_admin1.csv')

las = las%>%
  rename(adm1 = adminName)%>%
  mutate(date = as.Date(date, format="%d/%m/%Y"),
         month = month(date))%>%
  group_by(year, month, epiWeek, adm1)%>%
  summarize(num_cases = sum(value))%>%
  ungroup()

las$disease = 'Lassa fever'
las$country = 'Nigeria'

# las = las%>%
#   mutate(dummy_date = as.Date(paste0(year, '-', month, '-01'), format="%Y-%m-%d"))


las = las%>%
  group_by(year, month, adm1, disease, country)%>%
  summarize(num_cases=sum(num_cases))
las$continent = 'Africa'

# combine lassa and hantavirus
out = rbind(df, las)

## adding in Bolivia ##
bo = read.csv('data/human/raw/hanta/Bolivia_Hanta_BHF.csv')
bo$adm1 = stringr::str_to_title(bo$adm1)
bo$disease = stringr::str_to_title(bo$disease)
bo$disease = stringr::str_trim(bo$disease)

out = rbind(out, bo)

write.csv(out, 'data/processed/case_data.csv', row.names = F)
#-------------------- Create summary dataset for climate data -------------------- 
df = read.csv('data/processed/case_data.csv')

df_sum = df%>%
  group_by(country, year, month, disease)%>%
  summarize(monthcases = sum(num_cases, na.rm=T))%>%
  group_by(country, disease)%>%
  summarize(minyear = min(year),
            maxyear = max(year),
            max_monthly_cases = max(monthcases, na.rm=T),
            total_cases = sum(monthcases,na.rm=T))%>%
  mutate(country = ifelse(country == 'USA', 'United States of America', country))

write.csv(df_sum, 'data/processed/country_list.csv', row.names = F)
